+++ src/controllers/public.js
@@ -1,0 +1,29 @@
+const BaseController = require('../core/baseController');
+
+class Public extends BaseController {
+    constructor() {
+        super({
+            name: 'root',
+            path: '/',
+            routes: [
+                { method: 'GET', path: '/', handler: 'main' },
+                { method: 'GET', path: 'live', handler: 'live' },
+                { method: 'GET', path: 'ready', handler: 'ready' },
+            ]
+        });
+    }
+
+    async main(ctx) {
+        ctx.body = { message: 'API' };
+    }
+
+    async live(ctx) {
+        ctx.body = { message: 'API' };
+    }
+
+    async ready(ctx) {
+        ctx.body = { message: 'API' };
+    }
+}
+
+module.exports = Public;
